Bot: split into modules
