+++ app/docker-compose.yml
@@ -73,7 +73,7 @@ services:
        gunicorn 
        --worker-class aiohttp.worker.GunicornWebWorker
        --bind 0.0.0.0:80
-       bioconda_utils.bot:init_app
+       bioconda_utils.bot.web:start
        #(disable_internal_celery=True)
     depends_on:
       - mq
@@ -92,7 +92,7 @@ services:
     command: >-
       celery worker
       -l info
-      -A bioconda_utils.bot
+      -A bioconda_utils.bot.worker
       --without-heartbeat
     deploy:
       mode: replicated
