+++ bioconda_utils/bot/events.py
@@ -1,0 +1,40 @@
+"""
+Github Events
+"""
+
+import logging
+import re
+
+import gidgethub.routing
+
+from .commands import command_routes
+
+logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
+event_routes = gidgethub.routing.Router()  # pylint: disable=invalid-name
+
+BOT_ALIAS_RE = re.compile(r'@bioconda[- ]?bot', re.IGNORECASE)
+
+
+@event_routes.register("issue_comment", action="created")
+async def comment_created(event, ghapi, *args, **_kwargs):
+    """Dispatches @bioconda-bot commands
+
+    This function watches for comments on issues. Lines starting with
+    an @mention of the bot are considered commands and dispatched.
+    """
+    commands = [
+        line.lower().split()[1:]
+        for line in event.data['comment']['body'].splitlines()
+        if BOT_ALIAS_RE.match(line)
+    ]
+    if not commands:
+        logger.info("No command in comment")
+    for cmd, *args in commands:
+        logger.info("Dispatching %s - %s", cmd, args)
+        await command_routes.dispatch(cmd, event, ghapi, *args)
+
+
+@event_routes.register("status")
+async def demo_using_status(_event, _ghapi):
+    """Test function for status events"""
+    logger.warning("Status - not handling")
